requirements.txt
datasets==2.11.0
dvc==2.51.0
fastapi==0.88.0
gunicorn==20.1.0
httpx==0.23.3
lightning-bolts==0.6.0.post1
lightning-cloud==0.5.32
lightning-utilities==0.8.0
numpy==1.24.1
ordered-set==4.1.0
pandas==1.5.3
Pillow==9.3.0
pytest==7.2.2
requests==2.28.1
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tqdm==4.65.0
uvicorn==0.21.1
